+++ CHANGELOG.md
@@ -13,6 +13,7 @@ See also
 - add svg and html for the queer flag
 - add functionality to add the queer flag to the twitter box
 - add click event for the flag
+- still no other functionality
 
 ## [1.0.0-rc.1](https://github.com/stephfuchs/tampermonkey-queer-wisdom-for-twitter/releases/tag/1.0.0-rc.1)