+++ CHANGELOG.md
@@ -8,15 +8,27 @@ See also
 
 ## [Unreleased]
 
+### Added
+
 - add interval to check every 5 min whether the flag was deleted by Twitter (happens some time)
+- add function to create the the tweet link with the wisdom.
+
+### Changed
+
+- changed the flag element from `div` to `a` to make it easier with the random text
 
 ## [1.0.0-rc.2](https://github.com/stephfuchs/tampermonkey-queer-wisdom-for-twitter/releases/tag/1.0.0-rc.2)
+
+- still no other functionality
+
+### Added
 
 - add svg and html for the queer flag
 - add functionality to add the queer flag to the twitter box
 - add click event for the flag
-- still no other functionality
 
 ## [1.0.0-rc.1](https://github.com/stephfuchs/tampermonkey-queer-wisdom-for-twitter/releases/tag/1.0.0-rc.1)
+
+### Added
 
 - initial add of the user script, read me, changelog and license